scraper/requirements.txt
selenium>=4.0.0
pillow>=9.0.0
pytesseract>=0.3.10
requests>=2.28.0
